Simplified the JSON-decoding of dependencies.
It seemed to me that the way the list of dependencies was being
encoded and decoded in Elm.Package.Description was doing work that
really ought to be done by --- and in the case of the other fields
already is done by --- ToJSON and FromJSON instances. So I stuck
the dependency list in a newtype and moved the relevant code.

I also removed `getElmVersion`, as it seemed to be completely
redundant --- the `instance FromJSON Constraint` takes care of
that.